feat(tree-grid): moving some code to grid api service #2530
